The BookBayes Module

BookBayes — Module

module BookBayes

Library of functions given in Algorithms for Decision Making

Module Index

BookBayes.BayesianNetwork

BookBayes.ExactInference

BookBayes.Factor

BookBayes.FactorTable

BookBayes.Variable

BookBayes.VariableElimination

BookBayes.assignments

BookBayes.condition

BookBayes.condition

BookBayes.in_scope

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.marginalize

BookBayes.normalize!

BookBayes.probability

BookBayes.variablenames

Detailed API

BookBayes.BayesianNetwork — Type

struct BayesianNetwork

Represents a Bayesian Network.

Fields:

vars::Vector{Variable}:

Vector of Variables contained in network

factors::Vector{Factor}:

Vector of Factors, one for each Variable in vars, representing conditional probability distribution over that variable given parents of var in network.

graph::SimpleDiGraph{Int64}:

Represents the structure of the network variables and implicitly defines independence relationships.

BookBayes.ExactInference — Type

struct ExactInference

Singleton type to pass instances of as first arg of infer to trigger direct inference method.

BookBayes.Factor — Type

struct Factor

A Type representing a joint or conditional probability distribution over the variables in vars.

Fields:

vars:

The variables represented by this distribution

table:

Probability table mapping Variable assignments (NamedTuple) => probability (Float64)

BookBayes.FactorTable — Type

const FactorTable = Dict{NamedTuple,Float64}

Type alias of a Dictionary mapping tuples of Variable name=>value assignments to probabilities.

BookBayes.Variable — Type

struct Variable

A Type representing a single discrete variable.

Params:

name::Symbol

name of Variable

m::Int:

number of possible (discrete) values this Variable can take

BookBayes.VariableElimination — Type

struct VariableElimination

Singleton type to pass instances of as first arg of infer to trigger variable elimination method.

ordering

Iterable over variable indices indicating the order over the factors to use.

Base.:* — Method

function Base.:*(ϕ::Factor, ψ::Factor)

Multiply two (joint and conditional) Factors to form a new (joint) Factor.

Base.rand — Method

function Base.rand(bn::BayesianNetwork)

Direct sample from a joint distribution given by bn.

Returns: NamedTuple of Variable name => value pairs sampled from bn.

Base.rand — Method

function Base.rand(φ::Factor)

Direct sample a discrete Factor given by φ

BookBayes.assignments — Method

function assignments(vars::AbstractVector{Variable})

Get an array of all possible variable=>value assignment tuples for Variables in vars.

BookBayes.condition — Method

function condition(φ::Factor, name, value)

Method for factor conditioning. Take a factor ϕ and return a new factor whose table entries are consistent with the variable named name having value value.

BookBayes.condition — Method

function condition(φ::Factor, evidence)

Method for factor conditioning. Take a factor ϕ and apply evidence in the form of a named tuple.

BookBayes.in_scope — Method

function in_scope(name, ϕ)

predicate function: true if any Variable v in ϕ has v.name == name (ie if name is a variable in ϕ)

Arguments

name::Symbol: Name of variable to test for presense of in Factor ϕ.

ϕ::Factor: Factor to test if variable named <name> is present in.

BookBayes.infer — Method

function infer(M::DirectSampling, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the direct sampling inference method to draw m samples from network consistent with evidence.

Fields:

M::DirectSampling

Singleton value to dispatch this version of infer

bn

BayesianNetwork specifying original distribution

query

Tuple of Symbols of Variable names to get joint distribution over.

evidence

Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(D::MvNormal, query, evidencevars, evidence)

Get the conditional normal distribution over the query variables, given the evidence variables

\[\left[ \begin{array}{c} \bf a \\ \bf b \end{array} \right] \sim \mathcal{N} \left( \left[ \begin{array}{c} \bf \mu_a \\ \bf \mu_b \end{array} \right] , \left[ \begin{array}{cc} \bf A & \bf C \\ \bf C^T & \bf B \end{array} \right] \right)\]

The conditional distribution is:

\[p(\textbf{a} | \textbf{b}) = \mathcal{N}( \bf a\ |\ \mu_{a|b},\ \Sigma_{a|b})\]

\[\boxed{\bf \mu_{a|b} = \mu_a + CB^{-1}(b - \mu_b)}\]

\[\boxed{\Sigma_{a|b} = A - CB^{-1}C^T}\]

Examples

Consider:

\[\left[ \begin{array}{c} x_1 \\ x_2 \end{array} \right] \sim \mathcal{N} \left( \left[ \begin{array}{c} 0 \\ 1 \end{array} \right] , \left[ \begin{array}{cc} 3 & 1 \\ 1 & 2 \end{array} \right] \right)\]

Then the conditional distribution for x₁ given x₂ = 2 is:

\[\mu_{x_1|x_2=2} = 0 + 1 \cdot 2^{-1} \cdot (2 - 1) = 0.5\]

\[\Sigma_{x_1|x_2=2} = 3 - 1 \cdot 2^{-1}\cdot 1 = 2.5\]

julia>  d = MvNormal([0.0,1.0], [3.0 1.0; 1.0 2.0])
d = FullNormal(
	dim: 2
	μ: [0.0, 1.0]
	Σ: [3.0 1.0; 1.0 2.0]
	)

# get conditional dist for x₁ given x₂ = 2.0
julia> infer(d, [1], [2], [2.0])
FullNormal(
dim: 1
μ: [0.5]
Σ: [2.5]
)

BookBayes.infer — Method

function infer(M::ExactInference, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses direct inference.

Fields:

M::ExactInference

Singleton value to use this version of infer

bn

BayesianNetwork specifying original distribution

query

Tuple of Symbols of Variable names to get joint distribution over.

evidence

Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(M::VariableElimination, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the sum-product variable elimination algorithm.

Fields:

M::VariableElimination

Singleton value to dispatch this version of infer

bn

BayesianNetwork specifying original distribution

query

Tuple of Symbols of Variable names to get joint distribution over.

evidence

Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.marginalize — Method

function marginalize(ϕ::Factor, name)

Integrate ϕ::Factor by Variable with name name to get new Factor with marginal distribution over remaining Variables.

BookBayes.normalize! — Method

function normalize!(φ::Factor)

Update φ::Factor so all values sum to 1.0 while maintaining relative value (so values represent probabilities).

BookBayes.probability — Method

function probability(bn::BayesianNetwork, assignment)

Calculate the probability of a given variable assignment given the bayesian network bn.

Arguments

bn: The BayesianNetwork describing the joint distribution

assignment: a NamedTuple of :Symbol=>Integer pairs giving the categorical index value of each variable.

Examples

julia> X = Variable(:x, 2);

julia> Y = Variable(:y, 2);

julia> Z = Variable(:z, 2);

julia> ϕx = Factor([X, Y, Z], FactorTable(
           (x=1, y=1, z=1) => 0.08,
           (x=1, y=1, z=2) => 0.31,
           (x=1, y=2, z=1) => 0.09,
           (x=1, y=2, z=2) => 0.37,
           (x=2, y=1, z=1) => 0.01,
           (x=2, y=1, z=2) => 0.05,
           (x=2, y=2, z=1) => 0.02,
           (x=2, y=2, z=2) => 0.07,
       ));

julia> ϕy = Factor([Y], FactorTable(
           (y=1,) => 0.20,
           (y=2,) => 0.80,
       ));

julia> ϕz = Factor([Z], FactorTable(
           (z=1,) => 0.40,
           (z=2,) => 0.60,
       ));

julia> g = SimpleDiGraph(3);

julia> add_edge!(g,1,3);

julia> add_edge!(g,2,3);

julia> bn = BayesianNetwork([Y, Z, X], [ϕy, ϕz, ϕ], g);

julia> assignment = (x=1, y=2, z=1);

julia> probability(bn, assignment)
0.028800000000000006

BookBayes.variablenames — Method

function variablenames(φ::Factor)

Get a vector of variable names (Symbols) for the Variables in a Factor