The BookBayes Module

BookBayes — Module

module BookBayes

Library of functions given in Algorithms for Decision Making

Module Index

BookBayes.Factor

BookBayes.FactorTable

BookBayes.Variable

BookBayes.assignments

BookBayes.namedtuple

BookBayes.normalize!

BookBayes.variablenames

Detailed API

BookBayes.Factor — Type

struct Factor
    vars::Vector{Variable}
    table::FactorTable
end

Data structure representing a discrete conditional probability distribution

BookBayes.FactorTable — Type

const FactorTable = Dict{NamedTuple,Float64}

BookBayes.Variable — Type

struct Variable
    name::Symbol
    m::Int # number of possible values
end

Data structure representing a single node in a discrete bayesian network

BookBayes.assignments — Method

function assignments(vars::AbstractVector{Variable})

Get a vector of every possible valid namedtuple corresponding to a possible valid (Integer (Categorical)) value assignment of the Variables in vars.

ARGS

EXAMPLES

(:a => 3, :b => 1, :c => 7)

BookBayes.namedtuple — Method

function namedtuple(n::Array{Symbol,1}, v::Tuple)

Create a NamedTuple of pairs of ::Symbol => ::Any

ARGS

n: Array of Symbols giving names of assignment variables

v: Tuple of values, one per name Symbol

BookBayes.normalize! — Method

normalize!(ϕ::Factor)

Normalize a factor ϕ, which divides all the entries in the factor by the same scalar so they sum to 1.

BookBayes.variablenames — Method

variablenames(ϕ::Factor) = [var.name for var in ϕ.vars]

Get Vector of the variable names of all variables in ϕ