The BookBayes Module

BookBayes — Module

module BookBayes

Library of functions given in Algorithms for Decision Making

Module Index

BookBayes.BayesianNetwork

BookBayes.DAG

BookBayes.DirectSampling

BookBayes.ExactInference

BookBayes.Factor

BookBayes.FactorTable

BookBayes.GibbsSampling

BookBayes.K2Search

BookBayes.LikelihoodWeightedSampling

BookBayes.LocalDirectedGraphSearch

BookBayes.Variable

BookBayes.VariableElimination

BookBayes.are_markov_equivalent

BookBayes.assignments

BookBayes.bayesian_score

BookBayes.bayesian_score_component

BookBayes.blanket

BookBayes.condition

BookBayes.condition

BookBayes.fit

BookBayes.fit

BookBayes.gaussian_kernel

BookBayes.gibbs_sample

BookBayes.in_scope

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.kernel_density_estimate

BookBayes.marginalize

BookBayes.normalize!

BookBayes.prior

BookBayes.probability

BookBayes.rand_graph_neighbor

BookBayes.statistics

BookBayes.sub2ind

BookBayes.update_gibbs_sample

BookBayes.variablenames

Detailed API

BookBayes.BayesianNetwork — Type

struct BayesianNetwork

Represents a Bayesian Network.

Fields:

vars::Vector{Variable}:

Vector of Variables contained in network

factors::Vector{Factor}:

Vector of Factors, one for each Variable in vars, representing conditional probability distribution over that variable given parents of var in network.

graph::SimpleDiGraph{Int64}:

Represents the structure of the network variables and implicitly defines independence relationships.

BookBayes.DAG — Type

const DAG = SimpleDiGraph

Data structure representing a directed acyclic graph

BookBayes.DirectSampling — Type

struct DirectSampling
    m # number of samples
end

BookBayes.ExactInference — Type

struct ExactInference

Singleton type to pass instances of as first arg of infer to trigger direct inference method.

BookBayes.Factor — Type

struct Factor

A Type representing a joint or conditional probability distribution over the variables in vars.

Fields:

vars:

The variables represented by this distribution

table:

Probability table mapping Variable assignments (NamedTuple) => probability (Float64)

BookBayes.FactorTable — Type

const FactorTable = Dict{NamedTuple,Float64}

Type alias of a Dictionary mapping tuples of Variable name=>value assignments to probabilities.

BookBayes.GibbsSampling — Type

struct GibbsSampling
    m_samples # number of samples to use
    m_burnin # number of samples to discard during burn-in
    m_skip # number of samples to skip for thinning
    ordering # array of variable indices
end

Data structure specifying parameters to use when Gibbs Sampling a BayesianNetwork

BookBayes.K2Search — Type

struct K2Search
    ordering::Vector{Int} # variable ordering
end

Data structure with parameters to inform acyclic directed graph fitting following K2 Search algorithm.

Fields

* ordering::Vector{Int}
    Enforces a topological sort on the potential graph search space.

BookBayes.LikelihoodWeightedSampling — Type

struct LikelihoodWeightedSampling
    m # number of samples
end

BookBayes.LocalDirectedGraphSearch — Type

struct LocalDirectedGraphSearch

Data structure with parameters to inform graph fit operation

BookBayes.Variable — Type

struct Variable

A Type representing a single discrete variable.

Params:

name::Symbol

name of Variable

m::Int:

number of possible (discrete) values this Variable can take

BookBayes.VariableElimination — Type

struct VariableElimination

Singleton type to pass instances of as first arg of infer to trigger variable elimination method.

ordering

Iterable over variable indices indicating the order over the factors to use.

Base.:* — Method

function Base.:*(ϕ::Factor, ψ::Factor)

Multiply two (joint and conditional) Factors to form a new (joint) Factor.

Base.rand — Method

function Base.rand(bn::BayesianNetwork)

Direct sample from a joint distribution given by bn.

Returns: NamedTuple of Variable name => value pairs sampled from bn.

Base.rand — Method

function Base.rand(φ::Factor)

Direct sample a discrete Factor given by φ

BookBayes.are_markov_equivalent — Method

function are_markov_equivalent(G::DAG, H::DAG)

Determine whether the directed acyclic graphs G and H are Markov equivalent.

BookBayes.assignments — Method

function assignments(vars::AbstractVector{Variable})

Get an array of all possible variable=>value assignment tuples for Variables in vars.

BookBayes.bayesian_score — Method

function bayesian_score(vars, G, D)

Compute the Bayesian score for a list of variables vars and a graph G given data D. Uses a uniform prior αᵢⱼₖ = 1 for all i, j, and k.

BookBayes.bayesian_score_component — Method

function bayesian_score_component(M, α)

Compute the component of the Bayesian score for a single Variable.

Arguments

* M::Matrix{Integer}
    Count matrix of size qᵢ x rᵢ of data.

* α::Matrix{Union{Float, Integer}}
    Matrix (of same size as M) of Dirichlet priors of a single Variable's parameter.

BookBayes.blanket — Method

function blanket(bn::BayesianNetwork, a, i)

Calculate $P(X\\_{i} | x\\_{-i})$ from a Bayesian Network bn.

Arguments

bn:: BayesianNetwork

a::NamedTuple: Complete assignment of values to every variable in bn.

i::Integer Index of leave-one-out variable Xᵢ in network (according to some valid topological sort)

BookBayes.condition — Method

function condition(φ::Factor, name, value)

Method for factor conditioning. Take a factor ϕ and return a new factor whose table entries are consistent with the variable named name having value value.

BookBayes.condition — Method

function condition(φ::Factor, evidence)

Method for factor conditioning. Take a factor ϕ and apply evidence in the form of a named tuple.

BookBayes.fit — Method

function fit(method::K2Search, vars, D)

Find a semi-optimal graph structure over the Variables vars using the K2 Search algorithm given the dataset D.

Arguments

* method::K2Search
    Parameters to contol fit operation
* vars::Vector{Variables}
* D::Matrix{Integer}
    Data array of size n x m, where n == length(vars) and m == number of data points.

BookBayes.fit — Method

function fit(method::LocalDirectedGraphSearch, vars, D)

Fit a graph structure to data set D of samples of Variables vars, following Local Directed Graph Search algorithm.

BookBayes.gaussian_kernel — Method

gaussian_kernel(b)

Get a scalar probability density function of a zero-mean, b standard deviation gaussian.

BookBayes.gibbs_sample — Method

function gibbs_sample(a, bn, evidence, ordering, m)

Sample BayesianNetwork bn using Gibbs Sampling. Start from NamedTuple assignment a, given evidence, with ordering giving valid topological sort to follow, and take m samples.

Returns

Returns a new NamedTuple assignment approximately randomly sampled from bn given evidence.

BookBayes.in_scope — Method

function in_scope(name, ϕ)

predicate function: true if any Variable v in ϕ has v.name == name (ie if name is a variable in ϕ)

Arguments

name::Symbol: Name of variable to test for presense of in Factor ϕ.

ϕ::Factor: Factor to test if variable named <name> is present in.

BookBayes.infer — Method

function infer(M::DirectSampling, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the direct sampling inference method to draw m samples from network consistent with evidence.

Returns

Returns estimated joint distribution over query variables conditioned on evidence variable assignment.

Arguments

M::DirectSampling Singleton value to dispatch this version of infer

bn BayesianNetwork specifying original distribution

query Tuple of Symbols of Variable names to get joint distribution over.

evidence Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(D::MvNormal, query, evidencevars, evidence)

Get the conditional normal distribution over the query variables, given the evidence variables

\[\left[ \begin{array}{c} \bf a \\ \bf b \end{array} \right] \sim \mathcal{N} \left( \left[ \begin{array}{c} \bf \mu_a \\ \bf \mu_b \end{array} \right] , \left[ \begin{array}{cc} \bf A & \bf C \\ \bf C^T & \bf B \end{array} \right] \right)\]

The conditional distribution is:

\[p(\textbf{a} | \textbf{b}) = \mathcal{N}( \bf a\ |\ \mu_{a|b},\ \Sigma_{a|b})\]

\[\boxed{\bf \mu_{a|b} = \mu_a + CB^{-1}(b - \mu_b)}\]

\[\boxed{\Sigma_{a|b} = A - CB^{-1}C^T}\]

Examples

Consider:

\[\left[ \begin{array}{c} x_1 \\ x_2 \end{array} \right] \sim \mathcal{N} \left( \left[ \begin{array}{c} 0 \\ 1 \end{array} \right] , \left[ \begin{array}{cc} 3 & 1 \\ 1 & 2 \end{array} \right] \right)\]

Then the conditional distribution for x₁ given x₂ = 2 is:

\[\mu_{x_1|x_2=2} = 0 + 1 \cdot 2^{-1} \cdot (2 - 1) = 0.5\]

\[\Sigma_{x_1|x_2=2} = 3 - 1 \cdot 2^{-1}\cdot 1 = 2.5\]

julia>  d = MvNormal([0.0,1.0], [3.0 1.0; 1.0 2.0])
d = FullNormal(
	dim: 2
	μ: [0.0, 1.0]
	Σ: [3.0 1.0; 1.0 2.0]
	)

# get conditional dist for x₁ given x₂ = 2.0
julia> infer(d, [1], [2], [2.0])
FullNormal(
dim: 1
μ: [0.5]
Σ: [2.5]
)

BookBayes.infer — Method

function infer(M::ExactInference, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses direct inference.

Arguments

M::ExactInference

Singleton value to use this version of infer

bn::BayesianNetwork

BayesianNetwork specifying original distribution

query::Tuple

Tuple of Symbols of Variable names to get joint distribution over.

evidence::NamedTuple

Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(M::GibbsSampling, bn, query, evidence)

Estimate the conditional distribution over the query variables given the evidence values, using Gibbs Sampling

BookBayes.infer — Method

function infer(M::LikelihoodWeightedSampling, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the Likelihood Weighted Sampling inference method to draw m samples from network consistent with evidence.

Returns

Returns estimated joint distribution over query variables conditioned on evidence variable assignment.

Arguments

M::LikelihoodWeightedSampling Singleton value to dispatch this version of infer

bn BayesianNetwork specifying original distribution

query Tuple of Symbols of Variable names to get joint distribution over.

evidence Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(M::VariableElimination, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the sum-product variable elimination algorithm.

Fields:

M::VariableElimination

Singleton value to dispatch this version of infer

bn::BayesianNetwork

BayesianNetwork specifying original distribution

query::Tuple{Symbol}

Tuple of Symbols of Variable names to get joint distribution over.

evidence::NamedTuple

Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.kernel_density_estimate — Method

function kernel_density_estimate(φ, O)

BookBayes.marginalize — Method

function marginalize(ϕ::Factor, name)

Integrate ϕ::Factor by Variable with name name to get new Factor with marginal distribution over remaining Variables.

BookBayes.normalize! — Method

function normalize!(φ::Factor)

Update φ::Factor so all values sum to 1.0 while maintaining relative value (so values represent probabilities).

BookBayes.prior — Method

function prior(vars, G)

Generate a matrix of Dirichlet priors {αᵢⱼₖ = 1 ∀ i, j, k} corresponding to a uniform prior distribution over the parameters of a bayesian network.

Arguments

* vars::Vector{Variables}
    The Variables of the discrete BayesianNetwork
* G::DAG
    The bayesian directed acyclic graph structure

BookBayes.probability — Method

function probability(bn::BayesianNetwork, assignment)

Calculate the probability of a given variable assignment given the bayesian network bn.

Arguments

bn: The BayesianNetwork describing the joint distribution

assignment: a NamedTuple of :Symbol=>Integer pairs giving the categorical index value of each variable.

Examples

julia> X = Variable(:x, 2);

julia> Y = Variable(:y, 2);

julia> Z = Variable(:z, 2);

julia> ϕx = Factor([X, Y, Z], FactorTable(
           (x=1, y=1, z=1) => 0.08,
           (x=1, y=1, z=2) => 0.31,
           (x=1, y=2, z=1) => 0.09,
           (x=1, y=2, z=2) => 0.37,
           (x=2, y=1, z=1) => 0.01,
           (x=2, y=1, z=2) => 0.05,
           (x=2, y=2, z=1) => 0.02,
           (x=2, y=2, z=2) => 0.07,
       ));

julia> ϕy = Factor([Y], FactorTable(
           (y=1,) => 0.20,
           (y=2,) => 0.80,
       ));

julia> ϕz = Factor([Z], FactorTable(
           (z=1,) => 0.40,
           (z=2,) => 0.60,
       ));

julia> g = SimpleDiGraph(3);

julia> add_edge!(g,1,3);

julia> add_edge!(g,2,3);

julia> bn = BayesianNetwork([Y, Z, X], [ϕy, ϕz, ϕ], g);

julia> assignment = (x=1, y=2, z=1);

julia> probability(bn, assignment)
0.028800000000000006

BookBayes.rand_graph_neighbor — Method

function rand_graph_neighbor(G::DAG)

Generate a new acyclic graph that is a random neighbor of G.

BookBayes.statistics — Method

function statistics(vars, G, D::Matrix{Int})

Extract the statistics (ie counts) from a discrete dataset D assuming a Bayesian network with variables vars and structure G. The dataset is an n x m matrix, where n is the number of variables, and m is the number of data points.

Returns

Returns an array M of length n. The ith component consists of a qᵢ x rᵢ matrix of counts.

BookBayes.sub2ind — Method

function sub2ind(siz, x)

Convert a length n cartesian coordinate for indexing an n dimensional array into its linear (scalar) equivalent.

Arguments

siz::NTuple{Int}

Result of calling size(A) on the matrix of interest A.

x::NTuple{Int}

length(siz)-length cartesian index to convert from

BookBayes.update_gibbs_sample — Method

function update_gibbs_sample(a, bn, evidence, ordering)

Single Gibbs Sampling sample loop. Starting from complete assignment a of values to all variable in bn, sample each distribution $P(X\\_{i} | x\\_{-i})$ one at a time, replacing its entry in a with sampled value after each iteration.

Arguments

a::NamedTuple

bn::BayesianNetwork

evidence::NamedTuple

ordering::Vector{Integer}: Vector of indices of each variable in bayesian network according to valid topological sort.

Returns

Returns NamedTuple assignment updated with new sampled values

BookBayes.variablenames — Method

function variablenames(φ::Factor)

Get a vector of variable names (Symbols) for the Variables in a Factor