The BookBayes Module

BookBayes — Module

module BookBayes

Library of functions given in Algorithms for Decision Making

Module Index

BookBayes.ApproximateValueIteration

BookBayes.BayesianNetwork

BookBayes.BranchAndBound

BookBayes.CrossEntropyPolicySearch

BookBayes.DAG

BookBayes.DirectSampling

BookBayes.EvolutionStrategies

BookBayes.ExactInference

BookBayes.Factor

BookBayes.FactorTable

BookBayes.ForwardSearch

BookBayes.GaussSeidelValueIteration

BookBayes.GeneticPolicySearch

BookBayes.GibbsSampling

BookBayes.HeuristicSearch

BookBayes.HeuristicSearch

BookBayes.HookeJeevesPolicySearch

BookBayes.IsotropicEvolutionStrategies

BookBayes.K2Search

BookBayes.LabeledHeuristicSearch

BookBayes.LabeledHeuristicSearch

BookBayes.LikelihoodWeightedSampling

BookBayes.LinearProgramFormulation

BookBayes.LinearQuadraticProblem

BookBayes.LinearRegressionValueFunction

BookBayes.LinearRegressionValueFunction

BookBayes.LocalDirectedGraphSearch

BookBayes.LocallyWeightedValueFunction

BookBayes.LocallyWeightedValueFunction

BookBayes.MDP

BookBayes.MonteCarloPolicyEvaluation

BookBayes.MonteCarloPolicyEvaluation

BookBayes.MonteCarloPolicyEvaluation

BookBayes.MonteCarloTreeSearch

BookBayes.MonteCarloTreeSearch

BookBayes.MultilinearValueFunction

BookBayes.MultilinearValueFunction

BookBayes.NearestNeighborValueFunction

BookBayes.NearestNeighborValueFunction

BookBayes.PolicyIteration

BookBayes.RolloutLookahead

BookBayes.RolloutLookahead

BookBayes.SimpleProblem

BookBayes.SimplexValueFunction

BookBayes.SimplexValueFunction

BookBayes.SparseSampling

BookBayes.ValueFunctionPolicy

BookBayes.ValueFunctionPolicy

BookBayes.ValueIteration

BookBayes.Variable

BookBayes.VariableElimination

BookBayes.are_markov_equivalent

BookBayes.assignments

BookBayes.backup

BookBayes.bayesian_score

BookBayes.bayesian_score_component

BookBayes.blanket

BookBayes.bonus

BookBayes.branch_and_bound

BookBayes.condition

BookBayes.condition

BookBayes.evolution_strategy_weights

BookBayes.expand

BookBayes.explore

BookBayes.fit

BookBayes.fit

BookBayes.fit!

BookBayes.fit!

BookBayes.fit!

BookBayes.fit!

BookBayes.fit!

BookBayes.forward_search

BookBayes.gaussian_kernel

BookBayes.gibbs_sample

BookBayes.greedy

BookBayes.in_scope

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.infer

BookBayes.iterative_policy_evaluation

BookBayes.kernel_density_estimate

BookBayes.label!

BookBayes.lookahead

BookBayes.lookahead

BookBayes.marginalize

BookBayes.normalize!

BookBayes.optimize

BookBayes.optimize

BookBayes.optimize

BookBayes.optimize_dist

BookBayes.optimize_dist

BookBayes.optimize_dist

BookBayes.policy_evaluation

BookBayes.prior

BookBayes.probability

BookBayes.rand_graph_neighbor

BookBayes.randstep

BookBayes.rollout

BookBayes.simulate!

BookBayes.simulate!

BookBayes.simulate!

BookBayes.solve

BookBayes.solve

BookBayes.solve

BookBayes.solve

BookBayes.solve

BookBayes.solve

BookBayes.solve

BookBayes.solve

BookBayes.sparse_sampling

BookBayes.statistics

BookBayes.sub2ind

BookBayes.tensorform

BookBayes.update_gibbs_sample

BookBayes.value_of_information

BookBayes.variablenames

Detailed API

BookBayes.ApproximateValueIteration — Type

struct ApproximateValueIteration
    Uθ # initial parameterized value function that supports fit!
    S # set of discrete states for performing backups
    k_max # maximum number of iterations
end

BookBayes.BayesianNetwork — Type

struct BayesianNetwork

Represents a Bayesian Network.

Fields:

vars::Vector{Variable}:

Vector of Variables contained in network

factors::Vector{Factor}:

Vector of Factors, one for each Variable in vars, representing conditional probability distribution over that variable given parents of var in network.

graph::SimpleDiGraph{Int64}:

Represents the structure of the network variables and implicitly defines independence relationships.

BookBayes.BranchAndBound — Type

struct BranchAndBound
    𝒫 # problem
    d # depth
    Ulo # lower bound on value function at depth d Qhi # upper bound on action value function
end

BookBayes.CrossEntropyPolicySearch — Type

struct CrossEntropyPolicySearch
    p # initial distribution
    m # number of samples
    m_elite # number of elite samples
    k_max # number of iterations
end

BookBayes.DAG — Type

const DAG = SimpleDiGraph

Data structure representing a directed acyclic graph

BookBayes.DirectSampling — Type

struct DirectSampling
    m # number of samples
end

BookBayes.EvolutionStrategies — Type

struct EvolutionStrategies
    D       # distribution constructor
    ψ       # initial distribution parameterization
    ∇logp   # log search likelihood gradient
            #   note: ∇ψlogp = ∇logp(ψ, θ): ∇logp is a function of (ψ, θ)
    m       # number of samples
    α       # step factor
    k_max   # number of iterations
end

BookBayes.ExactInference — Type

struct ExactInference

Singleton type to pass instances of as first arg of infer to trigger direct inference method.

BookBayes.Factor — Type

struct Factor

A Type representing a joint or conditional probability distribution over the variables in vars.

Fields:

vars:

The variables represented by this distribution

table:

Probability table mapping Variable assignments (NamedTuple) => probability (Float64)

BookBayes.FactorTable — Type

const FactorTable = Dict{NamedTuple,Float64}

Type alias of a Dictionary mapping tuples of Variable name=>value assignments to probabilities.

BookBayes.ForwardSearch — Type

struct ForwardSearch
    𝒫 # problem
    d # depth
    U # value function at depth d
end

BookBayes.GaussSeidelValueIteration — Type

struct GaussSeidelValueIteration
    k_max # maximum number of iterations
end

BookBayes.GeneticPolicySearch — Type

struct GeneticPolicySearch
    θs # initial population
    σ # initial standard devidation
    m_elite # number of elite samples
    k_max # number of iterations
end

BookBayes.GibbsSampling — Type

struct GibbsSampling
    m_samples # number of samples to use
    m_burnin # number of samples to discard during burn-in
    m_skip # number of samples to skip for thinning
    ordering # array of variable indices
end

Data structure specifying parameters to use when Gibbs Sampling a BayesianNetwork

BookBayes.HeuristicSearch — Type

struct HeuristicSearch
    𝒫 # problem
    Uhi # upper bound on value function
    d # depth
    m # number of simulations
end

BookBayes.HeuristicSearch — Method

function (π::HeuristicSearch)(s)

Guranteed to converge to optimal value function iff Uhi is indeed an upperbound on U.

Complexity: 𝒪(m × d × |𝒮|×|𝒜|)

BookBayes.HookeJeevesPolicySearch — Type

struct HookeJeevesPolicySearch
    θ # initial parameterization
    α # step size
    c # step size reduction factor
    ε # termination step size
end

BookBayes.IsotropicEvolutionStrategies — Type

struct IsotropicEvolutionStrategies
    ψ # initial mean
    σ # initial standard devidation
    m # number of samples
    α # step factor
    k_max # number of iterations
end

BookBayes.K2Search — Type

struct K2Search
    ordering::Vector{Int} # variable ordering
end

Data structure with parameters to inform acyclic directed graph fitting following K2 Search algorithm.

Fields

* ordering::Vector{Int}
    Enforces a topological sort on the potential graph search space.

BookBayes.LabeledHeuristicSearch — Type

struct LabeledHeuristicSearch
    𝒫 # problem
    Uhi # upper bound on value function
    d # depth
    δ # gap threshold
end

BookBayes.LabeledHeuristicSearch — Method

function (π::LabeledHeuristicSearch)(s)

BookBayes.LikelihoodWeightedSampling — Type

struct LikelihoodWeightedSampling
    m # number of samples
end

BookBayes.LinearProgramFormulation — Type

struct LinearProgramFormulation end

BookBayes.LinearQuadraticProblem — Type

struct LinearQuadraticProblem
    Ts # transition matrix with respect to state
    Ta # transition matrix with respect to action
    Rs # reward matrix with respect to state (negative semidefinite)
    Ra # reward matrix with respect to action (negative definite)
    h_max # horizon
end

BookBayes.LinearRegressionValueFunction — Type

mutable struct LinearRegressionValueFunction
    β # basis vector function
    θ # vector of paramters
end

BookBayes.LinearRegressionValueFunction — Method

function (Uθ::LinearRegressionValueFunction)(s)

BookBayes.LocalDirectedGraphSearch — Type

struct LocalDirectedGraphSearch
    G::DAG          # initial graph
    k_max::Integer  # number of iterations
end

Data structure with parameters to inform graph fit operation.

BookBayes.LocallyWeightedValueFunction — Type

mutable struct LocallyWeightedValueFunction
    k # kernel function k(s, s′)
    S # set of discrete states
    θ # vector of values at states in S
end

BookBayes.LocallyWeightedValueFunction — Method

function (Uθ::LocallyWeightedValueFunction)(s)

BookBayes.MDP — Type

struct MDP
    γ # discount factor
    𝒮 # state space
    𝒜 # action space
    T # transition function
    R # reward function
    TR # sample transition and reward
end

BookBayes.MonteCarloPolicyEvaluation — Type

struct MonteCarloPolicyEvaluation
    𝒫 # problem
    b # initial state distribution
    d # depth
    m # number of samples
end

BookBayes.MonteCarloPolicyEvaluation — Method

(U::MonteCarloPolicyEvaluation)(π, θ) = U(s->π(θ, s))

Parameterized version where policy adjusted by parameter vector θ.

Returns: Ũ(π₍θ₎)

Complexity: 𝒪(m × d)

BookBayes.MonteCarloPolicyEvaluation — Method

function (U::MonteCarloPolicyEvaluation)(π)

Return the mean discounted reward of m random trajectories, each calculated by first sampling the initial state distribution, and then executing rollout from that state, using policy π.

\[U(\pi) = E_{\tau}[R(\tau)] = \sum_{\tau} p(\tau)R(\tau) \approx \frac{1}{m} \sum_{i=1}^{m} R(\tau^{(i)})\]

Returns: Ũ(π)

Complexity: 𝒪(m × d)

BookBayes.MonteCarloTreeSearch — Type

struct MonteCarloTreeSearch
    𝒫 # problem
    N # visit counts
    Q # action value estimates
    d # depth
    m # number of simulations
    c # exploration constant
    π # rollout policy
end

BookBayes.MonteCarloTreeSearch — Method

function (π::MonteCarloTreeSearch)(s)

Estimate optimal next action by first simulating m random trajectories. Each simulation improves estimated action-value function. The returned action is the optimal action over the estimated action-value function from the given state s.

Complexity: 𝒪(|𝒜| × d × m)

BookBayes.MultilinearValueFunction — Type

mutable struct MultilinearValueFunction
    o # position of lower-left corner
    δ # vector of widths
    θ # vector of values at states in S
end

BookBayes.MultilinearValueFunction — Method

function (Uθ::MultilinearValueFunction)(s)

Use multilinear interpolation to estimate the value of state vector s for known state values θ over a grid defined by lower-left vertex o and vector of widths δ. Verticies of the grid can all be written as o + δ.*i for some non-negative integer vector i.

BookBayes.NearestNeighborValueFunction — Type

mutable struct NearestNeighborValueFunction
    k # number of neighbors
    d # distance function d(s, s′)
    S # set of discrete states
    θ # vector of values at states in S
end

BookBayes.NearestNeighborValueFunction — Method

function (Uθ::NearestNeighborValueFunction)(s)

Approximate value function at s as average of value function over k nearest states (whose value function value is already known).

BookBayes.PolicyIteration — Type

struct PolicyIteration
    π # initial policy
    k_max # maximum number of iterations
end

BookBayes.RolloutLookahead — Type

struct RolloutLookahead
    𝒫 # problem
    π # rollout policy
    d # depth
end

BookBayes.RolloutLookahead — Method

function (π::RolloutLookahead)(s)

Policy function: similar to ValueFunctionPolicy but using a single call to rollout for value function.

Complexity: O(|𝒜|×|𝒮| × d)

Consider variation that uses average of m rollouts to calculate U(s) (vs just 1 with this version).

BookBayes.SimpleProblem — Type

struct SimpleProblem
    bn::BayesianNetwork
    chance_vars::Vector{Variable}
    decision_vars::Vector{Variable}
    utility_vars::Vector{Variable}
    utilities::Dict{Symbol, Vector{Float64}}
end

Data structure representing a decision network.

Fields

bn::BayesianNetwork

chance_vars::Vector{Variable}

decision_vars::Vector{Variable}

utility_vars::Vector{Variable}

utilities::Dict{Symbol, Vector{Float64}}

BookBayes.SimplexValueFunction — Type

mutable struct SimplexValueFunction
    o # position of lower-left corner
    δ # vector of widths
    θ # vector of values at states in S
end

BookBayes.SimplexValueFunction — Method

function (Uθ::SimplexValueFunction)(s)

Estimate value of state vector s for known values θ over a grid defined by a lower-left vertex o and vector of widths δ. Vertices of the grid can all be written o + δ.*i for some non-negative integral vector i.

BookBayes.SparseSampling — Type

struct SparseSampling
    𝒫 # problem
    d # depth
    m # number of samples
    U # value function at depth d
end

BookBayes.ValueFunctionPolicy — Type

struct ValueFunctionPolicy
    𝒫 # problem
    U # utility function
end

BookBayes.ValueFunctionPolicy — Method

function (π::ValueFunctionPolicy)(s)

Greedy policy function mapping states to greedy action.

BookBayes.ValueIteration — Type

struct ValueIteration
    k_max # maximum number of iterations
end

BookBayes.Variable — Type

struct Variable

A Type representing a single discrete variable.

Params:

name::Symbol

name of Variable

m::Int:

number of possible (discrete) values this Variable can take

BookBayes.VariableElimination — Type

struct VariableElimination

Singleton type to pass instances of as first arg of infer to trigger variable elimination method.

ordering

Iterable over variable indices indicating the order over the factors to use.

Base.:* — Method

function Base.:*(ϕ::Factor, ψ::Factor)

Multiply two (joint and conditional) Factors to form a new (joint) Factor.

Base.rand — Method

function Base.rand(bn::BayesianNetwork)

Direct sample from a joint distribution given by bn.

Returns: NamedTuple of Variable name => value pairs sampled from bn.

Base.rand — Method

function Base.rand(φ::Factor)

Direct sample a discrete Factor given by φ

BookBayes.are_markov_equivalent — Method

function are_markov_equivalent(G::DAG, H::DAG)

Determine whether the directed acyclic graphs G and H are Markov equivalent.

BookBayes.assignments — Method

function assignments(vars::AbstractVector{Variable})

Get an array of all possible variable=>value assignment tuples for Variables in vars.

BookBayes.backup — Method

function backup(𝒫::MDP, U, s)

Calculate maxₐ Q(s,a). Complexity: O(|𝒮|×|𝒜|).

return maximum(lookahead(𝒫, U, s, a) for a in 𝒫.𝒜)

BookBayes.bayesian_score — Method

function bayesian_score(vars, G, D)

Compute the Bayesian score for a list of variables vars and a graph G given data D. Uses a uniform prior αᵢⱼₖ = 1 for all i, j, and k.

BookBayes.bayesian_score_component — Method

function bayesian_score_component(M, α)

Compute the component of the Bayesian score for a single Variable.

Arguments

* M::Matrix{Integer}
    Count matrix of size qᵢ x rᵢ of data.

* α::Matrix{Union{Float, Integer}}
    Matrix (of same size as M) of Dirichlet priors of a single Variable's parameter.

BookBayes.blanket — Method

function blanket(bn::BayesianNetwork, a, i)

Calculate $P(X\\_{i} | x\\_{-i})$ from a Bayesian Network bn.

Arguments

bn:: BayesianNetwork

a::NamedTuple: Complete assignment of values to every variable in bn.

i::Integer Index of leave-one-out variable Xᵢ in network (according to some valid topological sort)

BookBayes.bonus — Method

bonus(Nsa, Ns)

Monte Carlo exploration bonus term helper function.

Complexity: 𝒪(1)

BookBayes.branch_and_bound — Method

function branch_and_bound(𝒫, s, d, Ulo, Qhi)

BookBayes.condition — Method

function condition(φ::Factor, name, value)

Method for factor conditioning. Take a factor ϕ and return a new factor whose table entries are consistent with the variable named name having value value.

BookBayes.condition — Method

function condition(φ::Factor, evidence)

Method for factor conditioning. Take a factor ϕ and apply evidence in the form of a named tuple.

BookBayes.evolution_strategy_weights — Method

function evolution_strategy_weights(m)

Complexity: 𝒪(m)

BookBayes.expand — Method

function expand(π::LabeledHeuristicSearch, U, solved, s)

BookBayes.explore — Method

function explore(π::MonteCarloTreeSearch, s)

Use monte carlo exploration heuristic to find next action to simulate. Balances need to explore state-action space with Complexity: 𝒪(|𝒜|)

BookBayes.fit! — Method

function fit!(Uθ::LinearRegressionValueFunction, S, U)

BookBayes.fit! — Method

function fit!(Uθ::LocallyWeightedValueFunction, S, U)

BookBayes.fit! — Method

function fit!(Uθ::MultilinearValueFunction, S, U)

BookBayes.fit! — Method

function fit!(Uθ::NearestNeighborValueFunction, S, U)

BookBayes.fit! — Method

function fit!(Uθ::SimplexValueFunction, S, U)

BookBayes.fit — Method

function fit(method::K2Search, vars, D)

Find a semi-optimal graph structure over the Variables vars using the K2 Search algorithm given the dataset D.

Arguments

* method::K2Search
    Parameters to contol fit operation
* vars::Vector{Variables}
* D::Matrix{Integer}
    Data array of size n x m, where n == length(vars) (ie number of Variables) and m == number of data points.

BookBayes.fit — Method

function fit(method::LocalDirectedGraphSearch, vars, D)

Fit a graph structure to data set D of samples of Variables vars, following Local Directed Graph Search algorithm.

BookBayes.forward_search — Method

function forward_search(𝒫, s, d, U)

Determine optimal action (and its value) to take from state s by expanding all possible transitions up to a depth d using depth-first search. U(s) used to evaluate the terminal (depth 0) value function.

Complexity: 𝒪( (|𝒜|×|𝒮|)ᵈ )

BookBayes.gaussian_kernel — Method

gaussian_kernel(b)

Get a scalar probability density function of a zero-mean, b standard deviation gaussian.

BookBayes.gibbs_sample — Method

function gibbs_sample(a, bn, evidence, ordering, m)

Sample BayesianNetwork bn using Gibbs Sampling. Start from NamedTuple assignment a, given evidence, with ordering giving valid topological sort to follow, and take m samples.

Returns

Returns a new NamedTuple assignment approximately randomly sampled from bn given evidence.

BookBayes.greedy — Method

function greedy(𝒫::MDP, U, s)

Find the greedy action and its expected utility starting from state s and using estimated optimal (greedy) state value function U. Calculates estimate as maximum over action space of 1-step lookahead:

u, a = _findmax(a->lookahead(𝒫, U, s, a), 𝒫.𝒜)

BookBayes.in_scope — Method

function in_scope(name, ϕ)

predicate function: true if any Variable v in ϕ has v.name == name (ie if name is a variable in ϕ)

Arguments

name::Symbol: Name of variable to test for presense of in Factor ϕ.

ϕ::Factor: Factor to test if variable named <name> is present in.

BookBayes.infer — Method

function infer(M::DirectSampling, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the direct sampling inference method to draw m samples from network consistent with evidence.

Returns

Returns estimated joint distribution over query variables conditioned on evidence variable assignment.

Arguments

M::DirectSampling Singleton value to dispatch this version of infer

bn BayesianNetwork specifying original distribution

query Tuple of Symbols of Variable names to get joint distribution over.

evidence Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(D::MvNormal, query, evidencevars, evidence)

Get the conditional normal distribution over the query variables, given the evidence variables

\[\left[ \begin{array}{c} \bf a \\ \bf b \end{array} \right] \sim \mathcal{N} \left( \left[ \begin{array}{c} \bf \mu_a \\ \bf \mu_b \end{array} \right] , \left[ \begin{array}{cc} \bf A & \bf C \\ \bf C^T & \bf B \end{array} \right] \right)\]

The conditional distribution is:

\[p(\textbf{a} | \textbf{b}) = \mathcal{N}( \bf a\ |\ \mu_{a|b},\ \Sigma_{a|b})\]

\[\boxed{\bf \mu_{a|b} = \mu_a + CB^{-1}(b - \mu_b)}\]

\[\boxed{\Sigma_{a|b} = A - CB^{-1}C^T}\]

Examples

Consider:

\[\left[ \begin{array}{c} x_1 \\ x_2 \end{array} \right] \sim \mathcal{N} \left( \left[ \begin{array}{c} 0 \\ 1 \end{array} \right] , \left[ \begin{array}{cc} 3 & 1 \\ 1 & 2 \end{array} \right] \right)\]

Then the conditional distribution for x₁ given x₂ = 2 is:

\[\mu_{x_1|x_2=2} = 0 + 1 \cdot 2^{-1} \cdot (2 - 1) = 0.5\]

\[\Sigma_{x_1|x_2=2} = 3 - 1 \cdot 2^{-1}\cdot 1 = 2.5\]

julia>  d = MvNormal([0.0,1.0], [3.0 1.0; 1.0 2.0])
d = FullNormal(
	dim: 2
	μ: [0.0, 1.0]
	Σ: [3.0 1.0; 1.0 2.0]
	)

# get conditional dist for x₁ given x₂ = 2.0
julia> infer(d, [1], [2], [2.0])
FullNormal(
dim: 1
μ: [0.5]
Σ: [2.5]
)

BookBayes.infer — Method

function infer(M::ExactInference, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses direct inference.

Arguments

M::ExactInference

Singleton value to use this version of infer

bn::BayesianNetwork

BayesianNetwork specifying original distribution

query::Tuple

Tuple of Symbols of Variable names to get joint distribution over.

evidence::NamedTuple

Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(M::GibbsSampling, bn, query, evidence)

Estimate the conditional distribution over the query variables given the evidence values, using Gibbs Sampling

BookBayes.infer — Method

function infer(M::LikelihoodWeightedSampling, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the Likelihood Weighted Sampling inference method to draw m samples from network consistent with evidence.

Returns

Returns estimated joint distribution over query variables conditioned on evidence variable assignment.

Arguments

M::LikelihoodWeightedSampling Singleton value to dispatch this version of infer

bn BayesianNetwork specifying original distribution

query Tuple of Symbols of Variable names to get joint distribution over.

evidence Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.infer — Method

function infer(M::VariableElimination, bn, query, evidence)

Get a joint distribution over query variables, given evidence. Uses the sum-product variable elimination algorithm.

Fields:

M::VariableElimination

Singleton value to dispatch this version of infer

bn::BayesianNetwork

BayesianNetwork specifying original distribution

query::Tuple{Symbol}

Tuple of Symbols of Variable names to get joint distribution over.

evidence::NamedTuple

Dictionary of Variable Symbol name => Variable value pairs specifying known values.

BookBayes.iterative_policy_evaluation — Method

function iterative_policy_evaluation(𝒫::MDP, π, k_max)

Approximate policy evaluation for policy π, using k_max iterations of 1-step lookahead, starting from U(s) = 0.

BookBayes.kernel_density_estimate — Method

function kernel_density_estimate(φ, O)

BookBayes.label! — Method

function label!(π::LabeledHeuristicSearch, U, solved, s)

BookBayes.lookahead — Method

function lookahead(𝒫::MDP, U, s, a)

Calculate state-action value function Q(s,a) from U and valid state transition distribution T:

Returns Q(s,a) = R(s,a) + γsum(T(s,a,s′)U(s′) for s′ in 𝒮)

BookBayes.lookahead — Method

function lookahead(𝒫::MDP, U::Vector, s, a)

Calculate state-action value function Q(s,a) from U and valid state transition distribution T:

return R(s,a) + γ*sum(T(s,a,s′)*U[i] for (i,s′) in enumerate(𝒮))

BookBayes.marginalize — Method

function marginalize(ϕ::Factor, name)

Integrate ϕ::Factor by Variable with name name to get new Factor with marginal distribution over remaining Variables.

BookBayes.normalize! — Method

function normalize!(φ::Factor)

Update φ::Factor so all values sum to 1.0 while maintaining relative value (so values represent probabilities).

BookBayes.optimize — Method

function optimize(M, π, U)

Policy parameter optimization over space of parameter distributions. For when M is a policy search type that lets policy parameters θ to come from a sampled distribution p, whose parameters are directly optimized to maximize E[U(θ)], θ ∼ p.

Returns: mode(p)

Complexity: depends on M

BookBayes.optimize — Method

function optimize(M::GeneticPolicySearch, π, U)

Returns: optimized policy parameterizaion θ_best.

BookBayes.optimize — Method

function optimize(M::HookeJeevesPolicySearch, π, U)

Returns: optimized policy parameterization θ_best.

BookBayes.optimize_dist — Method

function optimize_dist(M::CrossEntropyPolicySearch, π, U)

Train policy parameter distribution p for parameterized policy π using approximate value function U.

Returns: optimized distribution p ('p̃')

Complexity: 𝒪(k_max × m × d)

BookBayes.optimize_dist — Method

function optimize_dist(M::EvolutionStrategies, π, U)

BookBayes.optimize_dist — Method

function optimize_dist(M::IsotropicEvolutionStrategies, π, U)

BookBayes.policy_evaluation — Method

function policy_evaluation(𝒫::MDP, π)

Exact policy evaluation using matrix-based Moore-Penrose pseudoinverse equation. Requires O(|𝒮|³) time.

Returns: [ U(s) ∀ s ∈ {𝒮} ]

BookBayes.prior — Method

function prior(vars, G)

Generate a matrix of Dirichlet priors {αᵢⱼₖ = 1 ∀ i, j, k} corresponding to a uniform prior distribution over the parameters of a bayesian network.

Arguments

* vars::Vector{Variables}
    The Variables of the discrete BayesianNetwork
* G::DAG
    The bayesian directed acyclic graph structure

BookBayes.probability — Method

function probability(bn::BayesianNetwork, assignment)

Calculate the probability of a given variable assignment given the bayesian network bn.

Arguments

bn: The BayesianNetwork describing the joint distribution

assignment: a NamedTuple of :Symbol=>Integer pairs giving the categorical index value of each variable.

Examples

julia> X = Variable(:x, 2);

julia> Y = Variable(:y, 2);

julia> Z = Variable(:z, 2);

julia> ϕx = Factor([X, Y, Z], FactorTable(
           (x=1, y=1, z=1) => 0.08,
           (x=1, y=1, z=2) => 0.31,
           (x=1, y=2, z=1) => 0.09,
           (x=1, y=2, z=2) => 0.37,
           (x=2, y=1, z=1) => 0.01,
           (x=2, y=1, z=2) => 0.05,
           (x=2, y=2, z=1) => 0.02,
           (x=2, y=2, z=2) => 0.07,
       ));

julia> ϕy = Factor([Y], FactorTable(
           (y=1,) => 0.20,
           (y=2,) => 0.80,
       ));

julia> ϕz = Factor([Z], FactorTable(
           (z=1,) => 0.40,
           (z=2,) => 0.60,
       ));

julia> g = SimpleDiGraph(3);

julia> add_edge!(g,1,3);

julia> add_edge!(g,2,3);

julia> bn = BayesianNetwork([Y, Z, X], [ϕy, ϕz, ϕ], g);

julia> assignment = (x=1, y=2, z=1);

julia> probability(bn, assignment)
0.028800000000000006

BookBayes.rand_graph_neighbor — Method

function rand_graph_neighbor(G::DAG)

Generate a new acyclic graph that is a random neighbor of G.

BookBayes.randstep — Method

randstep(𝒫::MDP, s, a) = 𝒫.TR(s, a)

BookBayes.rollout — Method

function rollout(𝒫, s, π, d)

Complexity: 𝒪(d)

BookBayes.simulate! — Function

function simulate!(π::MonteCarloTreeSearch, s, d=π.d)

Execute exploration-bonus policy for 1 random trajectory, updating system counters, and return sample trajectory's value.

Complexity: 𝒪(|𝒜| × d)

BookBayes.simulate! — Method

function simulate!(π::HeuristicSearch, U, s)

Simulate depth-d trajectory following greedy policy and updating U along the way.

Complexity: 𝒪(d × |𝒜|×|𝒮|)

BookBayes.simulate! — Method

function simulate!(π::LabeledHeuristicSearch, U, solved, s)

BookBayes.solve — Method

function solve(M::ApproximateValueIteration, 𝒫::MDP)

BookBayes.solve — Method

solve(𝒫::MDP) = solve(LinearProgramFormulation(), 𝒫)

Default MDP solver is the LinearProgramFormulation solver.

BookBayes.solve — Method

function solve(M::GaussSeidelValueIteration, 𝒫::MDP)

Same as Value Iteration but update U one state at a time. Convergence speed depends on ordering of states in 𝒮.

BookBayes.solve — Method

function solve(M::LinearProgramFormulation, 𝒫::MDP)

BookBayes.solve — Method

function solve(𝒫::LinearQuadraticProblem)

BookBayes.solve — Method

function solve(M::PolicyIteration, 𝒫::MDP)

Iteratively improves an initial policy π to obtain an optimal policy for an MDP 𝒫 with discrete state and action spaces. Returns improved policy.

BookBayes.solve — Method

function solve(𝒫::SimpleProblem, evidence, M)

Given a decision network 𝒫 and observed variable assignment evidence, determine the optimal action variable assignment and associated expected utility, using inference method M.

Arguments

𝒫::SimpleProblem
    Decision network
evidence
    Vector of NamedTuples mapping Symbol name of Variable to assignment of observed values of condition vars.
M
    Method to use for inference over utility vars given evidence

BookBayes.solve — Method

function solve(M::ValueIteration, 𝒫::MDP)

Iteratively improves a value function U to obtain an optimal policy for an MDP 𝒫 with discrete state and action spaces. Terminates after k_max iterations.

Order O(k_max × |𝒮| × |𝒜| × |𝒮|)

Returns

Optimal policy π

BookBayes.sparse_sampling — Method

function sparse_sampling(𝒫, s, d, m, U)

Complexity: 𝒪( (|𝒜| × m)ᵈ )

BookBayes.statistics — Method

function statistics(vars, G, D::Matrix{Int})

Extract the statistics (ie counts) from a discrete dataset D assuming a Bayesian network with variables vars and structure G. The dataset is an n x m matrix, where n is the number of variables, and m is the number of data points.

Returns

Returns an array M of length n. The ith component consists of a qᵢ x rᵢ matrix of counts.

BookBayes.sub2ind — Method

function sub2ind(siz, x)

Convert a length n cartesian coordinate for indexing an n dimensional array into its linear (scalar) equivalent.

Arguments

siz::NTuple{Int}

Result of calling size(A) on the matrix of interest A.

x::NTuple{Int}

length(siz)-length cartesian index to convert from

BookBayes.tensorform — Method

function tensorform(𝒫::MDP)

Convert an MDP into its tensor form, where the states and actions consist of integer indicies, the reward function is a matrix, and the transition function is a three-dimensional tensor.

BookBayes.update_gibbs_sample — Method

function update_gibbs_sample(a, bn, evidence, ordering)

Single Gibbs Sampling sample loop. Starting from complete assignment a of values to all variable in bn, sample each distribution $P(X\\_{i} | x\\_{-i})$ one at a time, replacing its entry in a with sampled value after each iteration.

Arguments

a::NamedTuple

bn::BayesianNetwork

evidence::NamedTuple

ordering::Vector{Integer}: Vector of indices of each variable in bayesian network according to valid topological sort.

Returns

Returns NamedTuple assignment updated with new sampled values

BookBayes.value_of_information — Method

function value_of_information(𝒫, query, evidence, M)

Determine the Value of Information (VOI) of query variables, given evidence assignment of (specific) values to already observed variables, using inference method M.

BookBayes.variablenames — Method

function variablenames(φ::Factor)

Get a vector of variable names (Symbols) for the Variables in a Factor